# backend/docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrueAdmin API",
    "version": "0.1.0",
    "description": "Generated from TrueAdmin controller attributes. Code annotations provide default interface metadata."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/admin/organization/client-users": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_ClientUserController_create",
        "summary": "新增用户端账号",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\ClientUserController@create",
          "permission": "system:client-user:create",
          "permissionMode": "single",
          "permissions": [
            "system:client-user:create"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_ClientUserController_list",
        "summary": "用户端账号列表",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\ClientUserController@list",
          "permission": "system:client-user:list",
          "permissionMode": "single",
          "permissions": [
            "system:client-user:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/organization/client-users/{id}": {
      "delete": {
        "operationId": "App_Module_System_Http_Admin_Controller_ClientUserController_delete",
        "summary": "删除用户端账号",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\ClientUserController@delete",
          "permission": "system:client-user:delete",
          "permissionMode": "single",
          "permissions": [
            "system:client-user:delete"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_ClientUserController_detail",
        "summary": "用户端账号详情",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\ClientUserController@detail",
          "permission": "system:client-user:detail",
          "permissionMode": "single",
          "permissions": [
            "system:client-user:detail"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "App_Module_System_Http_Admin_Controller_ClientUserController_update",
        "summary": "编辑用户端账号",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\ClientUserController@update",
          "permission": "system:client-user:update",
          "permissionMode": "single",
          "permissions": [
            "system:client-user:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/organization/client-users/{id}/disable": {
      "put": {
        "operationId": "App_Module_System_Http_Admin_Controller_ClientUserController_disable",
        "summary": "禁用用户端账号",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\ClientUserController@disable",
          "permission": "system:client-user:status",
          "permissionMode": "single",
          "permissions": [
            "system:client-user:status"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/organization/client-users/{id}/enable": {
      "put": {
        "operationId": "App_Module_System_Http_Admin_Controller_ClientUserController_enable",
        "summary": "启用用户端账号",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\ClientUserController@enable",
          "permission": "system:client-user:status",
          "permissionMode": "single",
          "permissions": [
            "system:client-user:status"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/organization/users": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminUserController_create",
        "summary": "新增管理员用户",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminUserController@create",
          "permission": "system:user:create",
          "permissionMode": "single",
          "permissions": [
            "system:user:create"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminUserController_list",
        "summary": "管理员用户列表",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminUserController@list",
          "permission": "system:user:list",
          "permissionMode": "single",
          "permissions": [
            "system:user:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/organization/users/{id}": {
      "delete": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminUserController_delete",
        "summary": "删除管理员用户",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminUserController@delete",
          "permission": "system:user:delete",
          "permissionMode": "single",
          "permissions": [
            "system:user:delete"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminUserController_detail",
        "summary": "管理员用户详情",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminUserController@detail",
          "permission": "system:user:detail",
          "permissionMode": "single",
          "permissions": [
            "system:user:detail"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminUserController_update",
        "summary": "编辑管理员用户",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminUserController@update",
          "permission": "system:user:update",
          "permissionMode": "single",
          "permissions": [
            "system:user:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}